Move custom javascript out of the packs directory so they aren't compiled as seperate entry points.
