+++ app/javascript/src/reports.js
